Modulated the student page
